css/css-borders-4/parsing/border-radius-left-valid.html

<!DOCTYPE html>
<html>
    <head>
        <meta charset="utf-8" />
        <title>CSS Borders 4 Test: Parsing 'border-radius-left' with valid values</title>
        <link rel="author" title="Sebastian Zartner" href="mailto:sebastianzartner@gmail.com" />
        <link rel="help" href="https://drafts.csswg.org/css-borders-4/#corner-sizing" />
        <meta name="assert" content="'border-radius-left' supports the syntax <length-percentage [0,∞]>{1,2} [ / <length-percentage [0,∞]>{1,2} ]?" />
		<script src="/resources/testharness.js"></script>
		<script src="/resources/testharnessreport.js"></script>
		<script src="/css/support/parsing-testcommon.js"></script>
	</head>
	<body>
		<script>
			test_valid_value("border-radius-left", "1px");
			test_valid_value("border-radius-left", "10%");
			test_valid_value("border-radius-left", "1px 5%");
			test_valid_value("border-radius-left", "5% 1px");
			test_valid_value("border-radius-left", "1px / 2px");
			test_valid_value("border-radius-left", "1px / 1px 2%");
			test_valid_value("border-radius-left", "1px 2% / 3%");
			test_valid_value("border-radius-left", "1px 2% / 3px 4px");
		</script>
	</body>
</html>
